Turtle -> OrigTurtle, LongTurtle -> Turtle
